requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
langchain-core>=0.1.7
langchain-huggingface>=0.0.1
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pydantic>=2.5.0
flask>=3.0.0
flask-cors>=4.0.0
google-generativeai>=0.3.2
pillow>=10.0.0
pypdf2>=3.0.0
python-docx>=1.1.0
gtts>=2.4.0
gunicorn>=21.2.0
transformers>=4.35.0
torch>=2.0.0
huggingface-hub>=0.17.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
